add statusToBadgeVariant
