+++ src/pages/namespace/Instances/List/Row.tsx
@@ -1,6 +1,4 @@
-import { ComponentProps, FC } from "react";
 import { Link, useNavigate } from "react-router-dom";
-import { Stats, stat } from "fs";
 import { TableCell, TableRow } from "~/design/Table";
 import {
   Tooltip,
@@ -10,26 +8,12 @@ import {
 } from "~/design/Tooltip";
 
 import Badge from "~/design/Badge";
+import { FC } from "react";
 import { InstanceSchemaType } from "~/api/instances/schema";
 import { pages } from "~/util/router/pages";
+import { statusToBadgeVariant } from "./utils";
+import { useTranslation } from "react-i18next";
 import useUpdatedAt from "~/hooksNext/useUpdatedAt";
-
-type BadgeVariant = ComponentProps<typeof Badge>["variant"];
-type InstanceStatus = InstanceSchemaType["status"];
-
-const statusToBadgeVariant = (status: InstanceStatus): BadgeVariant => {
-  switch (status) {
-    case "complete":
-      return "success";
-    case "crashed":
-    case "failed":
-      return "destructive";
-    case "pending":
-      return undefined;
-    default:
-      break;
-  }
-};
 
 const InstanceTableRow: FC<{
   instance: InstanceSchemaType;
@@ -39,6 +23,7 @@ const InstanceTableRow: FC<{
   const updatedAt = useUpdatedAt(instance.updatedAt);
   const createdAt = useUpdatedAt(instance.createdAt);
   const navigate = useNavigate();
+  const { t } = useTranslation();
 
   const isLatestRevision = revision === "latest";
 
@@ -61,7 +46,7 @@ const InstanceTableRow: FC<{
             <TooltipTrigger>
               <Link
                 onClick={(e) => {
-                  e.stopPropagation(); // prevent the onClick on the row from firing
+                  e.stopPropagation(); // prevent the onClick on the row from firing when clicking the workflow link
                 }}
                 to={pages.explorer.createHref({
                   namespace,
@@ -77,12 +62,7 @@ const InstanceTableRow: FC<{
             <TooltipContent>click to open workflow</TooltipContent>
           </Tooltip>
         </TableCell>
-        <TableCell
-          className="w-32"
-          onClick={(e) => {
-            e.stopPropagation();
-          }}
-        >
+        <TableCell className="w-32">
           <Badge variant="outline">{instance.id.slice(0, 8)}</Badge>
         </TableCell>
         <TableCell className="w-28">
